Super tight code. Tidied up project structure and modules. 	new file:   arduino_test/arduino/mmw_led_breath_purple/mmw_led_breath_purple.ino 	deleted:    clip_manager.py 	modified:   signifier.py 	new file:   signify/clip.py 	new file:   signify/clip_manager.py 	new file:   signify/clip_utils.py 	new file:   signify/sig_utils.py
