+++ signifier.py
@@ -1,5 +1,4 @@
 
-
 #    _________.__              .__       .__    .__              
 #   /   _____/|__| ____   ____ |__|_____ |  |__ |__| ___________ 
 #   \_____  \ |  |/ ___\ /    \|  \____ \|  |  \|  |/ __ \_  __ \
@@ -7,71 +6,20 @@
 #  /_______  /|__\___  /|___|  /__|   __/|___|  /__|\___  >__|   
 #          \/   /_____/      \/   |__|        \/        \/
 
-
-# Displaying audio hardware
+# Displaying audio hardware:
 # aplay -l
 
-# For resolving error 'ALSA lib pcm_direct.c:1846:(_snd_pcm_direct_get_slave_ipc_offset) Invalid value for card'
+# Create a virtual audio loopback device:
+# sudo modprobe snd-aloop
+
+# For resolving error 'ALSA lib pcm_direct.c:1846:(_snd_pcm_direct_get_slave_ipc_offset) Invalid value for card'??
 # sudo usermod -aG audio $USER
+# Or you know... hardcode? :(
 
 
-
-
-
-# Web UI monitoring/control schema:
-#
-# (*) indicates GUI-controllable variable
-#
-# Clip Manager:
-#   - Base Path (str)
-#   - Collections (list)
-#   - Active Collection (str) (*)
-#   - Default Volume (float) (*)
-#   - Default Fade (list: fade-down, fade-up) ms (*)
-#
-# Collection:
-#   - Title
-#   - Path
-#   - Audio Files (names) (list)
-#   - Active Clips (int) (*)
-#
-# Inactive Pool (clips) (list) (*):
-#   - Clip Objects (list)
-# Active Pool (clips) (list) (*)
-#   - Clip Objects (list)
-#
-# Clip Object:
-#   - Root (str)
-#   - Name (str) (filename)
-#   - Full path (str)
-#   - Length (float)
-#   - Category (str)
-#   - Looping (bool)
-#   - Channel (int)
-#   - Playing (bool) [play/stop] (*) (associated with inactive/active pools)
-#   - Volume (float) (*)
-#
-# Mixer:
-#   - Channels (list) (int)
-#   - Volume (float) (*)
-#   - Playing (bool) [play/stop] (*) (controls clip playing status)
-#
-# Composition Dynamics:
-#   - Number of Clips (int) (*)
-#   - Quiet Level (int) (*)
-#   - Busy Level (int) (*)
-#
-# Scheduler Timers:
-#   - Collection Timer (int) (*)
-#   - Clip Timer (int) (*)
-#   - Volume Timer (int) (*)
-
-
-
-import logging, os, random, signal, sys, time, schedule
+import logging, os, signal, sys, time, schedule
 import pygame as pg
-
-from clip_manager import ClipManager
+from signify.clip_manager import ClipManager
 
 # Initialise logging
 logging.basicConfig(level=logging.DEBUG)
@@ -80,9 +28,10 @@ logger.setLevel(logging.DEBUG)
 
 os.environ["SDL_VIDEODRIVER"] = "dummy"
 
+clip_manager = None
 CLIP_EVENT = pg.USEREVENT+1
 
-# TODO Create method to remotely manage this dictionary
+# TODO Export as JSON and create method to remotely manage this dictionary
 config = {
     'audio':
     {
@@ -91,81 +40,109 @@ config = {
         'bit_size':-16,
         'buffer':2048
     },
-    'composition':
+    'clip_manager':
     {
-        'quiet_level':3,
-        'busy_level':6,
-        'max_playtime':60,
-    },
-    'clips':
-    {
-        'base_path':'/home/pi/Signifier/audio',
         'fail_retry_delay':5,
+        'base_path':'/home/pi/Signifier/audio',
         'valid_extensions':['.wav'],
         'strict_distribution':False,
-        'collection_size':12,
         'volume':0.5,
         'fade_in':2000,
-        'fade_out':3000
+        'fade_out':3000,
+        'max_playtime':60,
+        'categories':
+        {
+            'oneshot':{'threshold':0, 'is_loop':False, 'loop_range':[0,0]},
+            'short':{'threshold':5, 'is_loop':False, 'loop_range':[2,6]},
+            'medium':{'threshold':10, 'is_loop':True, 'loop_range':[0,0]},
+            'loop':{'threshold':30, 'is_loop':True, 'loop_range':[0,0]}
+        }
     },
     'jobs':
     {
-        'collection':96,
-        'clips':16,
-        'volume':2
+        'collection':
+        {
+            'state':True,
+            'timer':96,
+            'parameters': {'size':12}
+        },
+        'composition':
+        {
+            'state':True,
+            'timer':16,
+            'parameters': {'quiet_level':3, 'busy_level':6}
+        },
+        'volume':
+        {
+            'state':True,
+            'timer':2,
+            'parameters': {'speed':5, 'weight':0}
+        }
     }
 }
 
-jobs = {'collection':96, 'clips':16, 'volume':2}
-clip_manager = None
 
 
-def update_clips():
-    print()
-    changed = set()
-    changed = clip_manager.check_finished()
-    if clip_manager.clips_playing() < config['composition']['quiet_level']:
-        changed.update(clip_manager.play_clip())
-    elif clip_manager.clips_playing() > config['composition']['busy_level']:
-        changed.update(clip_manager.stop_clip())
-    # Stop a random clip after a certain amount of time?
-    print()
+#  _________                __                .__   
+#  \_   ___ \  ____   _____/  |________  ____ |  |  
+#  /    \  \/ /  _ \ /    \   __\_  __ \/  _ \|  |  
+#  \     \___(  <_> )   |  \  |  |  | \(  <_> )  |__
+#   \______  /\____/|___|  /__|  |__|   \____/|____/
+#          \/            \/                         
 
-def modulate_volume():
-    clip_manager.modulate_volumes()
-
-def stop_random_clip():
-    clip_manager.stop_clip()
-
-
-def new_collection():
+def new_collection(pool_size=None):
+    pool_size = config['jobs']['collection']['parameters']['size'] if pool_size is None else pool_size
     print()
     set_jobs(False)
     pg.event.clear()
     stop_all_clips()
     while pg.mixer.get_busy():
         time.sleep(0.1)
-    
-    while (collection := clip_manager.select_collection("S06") is None):
-        logger.info(f'Trying another collection in {config["clips"]["fail_retry_delay"]} seconds...')
-        time.sleep(config["clips"]["fail_retry_delay"])
+    while (collection := clip_manager.select_collection() is None):
+        logger.info(f'Trying another collection in {config["clip_manager"]["fail_retry_delay"]} seconds...')
+        time.sleep(config["clip_manager"]["fail_retry_delay"])
         new_collection()
-
-    set_jobs(True)
     logger.info(f'NEW COLLECTION JOB DONE.')
     print()
+    time.sleep(2)
+    set_jobs(True)
     #update_clips()
     pg.event.clear()
-    
 
-def stop_all_clips(fade_time=config['clips']['fade_out']):
+def modulate_volume(speed=None, weight=None):
+    """Randomly modulate the Channel volumes for all Clip(s) in the active pool.\n 
+    - "speed=(int)" is the maximum volume jump per tick as a percentage of the total volume.
+    1 is slow, 10 is very quick.\n - "weight=(float)" is a signed normalised float (-1.0 to 1.0)
+    that weighs the random steps towards either direction."""
+    speed = config['jobs']['volume']['parameters']['speed'] if speed is None else speed
+    weight = config['jobs']['volume']['parameters']['weight'] if weight is None else weight
+    clip_manager.modulate_volumes(speed, weight)
+
+def update_clips(quiet_level=None, busy_level=None):
+    """Ensure the clip manager is playing an appropriate number of clips,
+    and tidy lingering completed clips from the active pool."""
+    quiet_level = config['jobs']['composition']['parameters']['quiet_level'] if quiet_level is None else quiet_level
+    busy_level = config['jobs']['composition']['parameters']['busy_level'] if busy_level is None else busy_level
+    print()
+    changed = set()
+    changed = clip_manager.check_finished()
+    if clip_manager.clips_playing() < quiet_level:
+        changed.update(clip_manager.play_clip())
+    elif clip_manager.clips_playing() > busy_level:
+        changed.update(clip_manager.stop_clip())
+    # Stop a random clip after a certain amount of time?
+    print()
+
+def stop_random_clip():
+    clip_manager.stop_clip()
+
+def stop_all_clips(fade_time=None):
     """Fadeout and stop all active clips"""
+    fade_time = config['clip_manager']['fade_out'] if fade_time is None else fade_time
     if pg.mixer.get_init():
         if pg.mixer.get_busy():
             logger.info(f'Stopping audio clips, with {fade_time}ms fade...')
             pg.mixer.fadeout(fade_time)
-
-
 
 def set_jobs(state=True):
     """Add/remove jobs that control playback modulation."""
@@ -175,10 +152,17 @@ def set_jobs(state=True):
         schedule.clear('modulation')
     if state:
         logger.debug(f'Scheduling modulation jobs...')
-        schedule.every(config['jobs']['clips']).seconds.do(update_clips).tag('modulation')
-        schedule.every(config['jobs']['volume']).seconds.do(modulate_volume).tag('modulation')
+        schedule.every(config['jobs']['composition']['timer']).seconds.do(update_clips).tag('modulation')
+        schedule.every(config['jobs']['volume']['timer']).seconds.do(modulate_volume).tag('modulation')
         logger.debug(f'Jobs: {schedule.get_jobs("modulation")}')
 
+
+#    _________       __                
+#   /   _____/ _____/  |_ __ ________  
+#   \_____  \_/ __ \   __\  |  \____ \ 
+#   /        \  ___/|  | |  |  /  |_> >
+#  /_______  /\___  >__| |____/|   __/ 
+#          \/     \/           |__|    
 
 def audio_device_check() -> str:
     # TODO Create functional mechanism to find and test audio devices
@@ -224,7 +208,7 @@ def prepare_playback_engine():
     pg.mixer.init()
     pg.init()
     logger.debug(f'Audio mixer successfully configured with device: "{device} {pg.mixer.get_init()}"')
-
+    print()
 
 class ExitHandler:
     signals = { signal.SIGINT:'SIGINT', signal.SIGTERM:'SIGTERM' }
@@ -246,6 +230,14 @@ class ExitHandler:
         sys.exit()
 
 
+
+#     _____         .__        
+#    /     \ _____  |__| ____  
+#   /  \ /  \\__  \ |  |/    \ 
+#  /    Y    \/ __ \|  |   |  \
+#  \____|__  (____  /__|___|  /
+#          \/     \/        \/ 
+
 if __name__ == '__main__':
     print()
     logger.info('Prepare to be Signified!!')
@@ -255,14 +247,14 @@ if __name__ == '__main__':
 
     prepare_playback_engine()
     try:
-        clip_manager = ClipManager(pg.mixer, config['clips'])
+        clip_manager = ClipManager(config['clip_manager'], pg.mixer, CLIP_EVENT)
     except OSError:
         exit_handler.shutdown()
 
     clip_manager.init_library()
 
     new_collection()
-    coll_job = schedule.every(config['jobs']['collection']).seconds.do(new_collection)
+    coll_job = schedule.every(config['jobs']['collection']['timer']).seconds.do(new_collection)
     #schedule.run_all()
 
     # Main loop
